More refactoring to break up all of the large reconciler files (#76)
