+++ CHANGELOG.md
@@ -1,3 +1,117 @@
+## [2026-01-12 22:15] - Phase 6: Bind9Instance Reconciler Modular Refactoring Complete
+
+**Author:** Erick Bourgeois
+
+### Changed
+- **[src/reconcilers/bind9instance/](src/reconcilers/bind9instance/)**: Refactored from single 1,252-line file into 6 focused modules (~1,304 lines total)
+  - **Created modular directory structure** following proven pattern from DNSZone and Bind9Cluster:
+    - `mod.rs` (~272 lines) - Main reconciliation orchestration + finalizer implementation + public API
+    - `resources.rs` (~501 lines) - Resource lifecycle (`ConfigMap`, Deployment, Service, Secret, ServiceAccount)
+    - `status_helpers.rs` (~244 lines) - Status computation from pod health and status patching
+    - `zones.rs` (~134 lines) - Zone reconciliation logic (reverse selector pattern)
+    - `cluster_helpers.rs` (~112 lines) - Cluster integration and reference management
+    - `types.rs` (~41 lines) - Shared type re-exports and imports
+  - **Extracted resource management** - All Kubernetes resource create/update/delete operations centralized
+  - **Separated status concerns** - Status calculation from deployment/pods isolated for testing
+  - **Isolated zone reconciliation** - Zone reference updates in dedicated module
+  - **Centralized cluster integration** - Cluster info fetching and reference building in focused module
+  - All 637 tests passing, zero clippy warnings
+
+### Why
+**Problem**: Single 1,252-line file with mixed concerns:
+- Resource lifecycle management (~402 lines) - ConfigMap, Deployment, Service, Secret, ServiceAccount
+- Status computation and updates (~247 lines) - Pod health, readiness, status patching
+- Zone reconciliation (~100 lines) - Updating instance status with zone references
+- Cluster integration (~147 lines) - Fetching cluster info, building references
+- Main reconciliation orchestration (~149 lines)
+- All concerns intermingled in one file
+
+**Solution**:
+- Applied proven modular extraction pattern from DNSZone (Phases 1-2) and Bind9Cluster (Phase 5)
+- Separated logical concerns into focused modules with clear responsibilities
+- Maintained backward compatibility through public API re-exports (`reconcile_instance_zones`)
+- Implemented finalizer cleanup trait in mod.rs for proper resource deletion
+- Used `pub(super)` visibility for internal helper functions
+
+**Benefits**:
+- **78.3% main file reduction** (1,252 → 272 lines orchestration + finalizer)
+- **Better Organization**: Clear separation of concerns across 6 focused modules
+- **Easier Testing**: Each module can be tested in isolation
+- **Improved Maintainability**: Changes are localized to specific modules (e.g., resource updates only touch resources.rs)
+- **Reduced Cognitive Load**: Each file focuses on one responsibility
+- **Follows Proven Pattern**: Same modular structure as DNSZone and Bind9Cluster refactoring
+- **Minimal Overhead**: Total codebase only ~52 lines more (4.2% increase for module declarations)
+
+### Impact
+- ✅ **Main orchestration** (mod.rs): 272 lines - Reconciliation workflow + finalizer + module exports
+- ✅ **Resource management** (resources.rs): 501 lines - Complete resource lifecycle (largest module)
+- ✅ **Status helpers** (status_helpers.rs): 244 lines - Status computation and patching
+- ✅ **Zone reconciliation** (zones.rs): 134 lines - Zone reference updates (reverse selector)
+- ✅ **Cluster integration** (cluster_helpers.rs): 112 lines - Cluster info and references
+- ✅ **Shared types** (types.rs): 41 lines - Common imports and re-exports
+- ✅ **Total: 1,304 lines** (vs 1,252 original = +52 lines module overhead, 4.2% increase)
+- ✅ **All 637 tests passing** (up from 594 in Phase 5 - additional tests added)
+- ✅ **Zero clippy warnings**
+- ✅ **Backward compatible** - Public API unchanged (`reconcile_instance_zones` re-exported)
+
+---
+
+## [2026-01-12 21:30] - Phase 5: Bind9Cluster Reconciler Modular Refactoring Complete
+
+**Author:** Erick Bourgeois
+
+### Changed
+- **[src/reconcilers/bind9cluster/](src/reconcilers/bind9cluster/)**: Refactored from single 1,488-line file into 6 focused modules (~1,600 lines total)
+  - **Created modular directory structure** following proven dnszone.rs pattern:
+    - `mod.rs` (~226 lines) - Main reconciliation orchestration + public API exports
+    - `instances.rs` (~901 lines) - Instance lifecycle management (create, update, delete)
+    - `status_helpers.rs` (~255 lines) - Status calculation and update helpers
+    - `drift.rs` (~96 lines) - Instance drift detection logic
+    - `config.rs` (~74 lines) - Cluster `ConfigMap` management
+    - `types.rs` (~48 lines) - Shared type re-exports and imports
+  - **Extracted instance management logic** - All instance creation, update, and deletion operations centralized in dedicated module
+  - **Separated status concerns** - Status calculation and patching logic isolated for easier testing
+  - **Isolated drift detection** - Drift detection between desired and actual state in focused module
+  - **Centralized config management** - `ConfigMap` creation and updates in dedicated module
+  - All 594 tests passing, zero clippy warnings
+
+### Why
+**Problem**: Single 1,488-line file with mixed concerns:
+- Instance lifecycle management (~750 lines)
+- Status calculation and updates (~230 lines)
+- Drift detection (~81 lines)
+- `ConfigMap` management (~56 lines)
+- All concerns intermingled in one file
+
+**Solution**:
+- Applied proven modular extraction pattern from DNSZone refactoring (Phases 1-2)
+- Separated logical concerns into focused modules with clear responsibilities
+- Maintained backward compatibility through careful re-exports
+- Followed Rust module conventions with `mod.rs` orchestration
+
+**Benefits**:
+- **89.9% main file reduction** (1,488 → 226 lines orchestration)
+- **Better Organization**: Clear separation of concerns across 6 focused modules
+- **Easier Testing**: Each module can be tested in isolation
+- **Improved Maintainability**: Changes are localized to specific modules
+- **Reduced Cognitive Load**: Each file focuses on one responsibility
+- **Follows Proven Pattern**: Same modular structure as DNSZone refactoring
+- **No Code Duplication**: Total codebase only ~112 lines more (module overhead)
+
+### Impact
+- ✅ **Main orchestration** (mod.rs): 226 lines - Reconciliation workflow + module exports
+- ✅ **Instance management** (instances.rs): 901 lines - Complete instance lifecycle
+- ✅ **Status helpers** (status_helpers.rs): 255 lines - Status computation and patching
+- ✅ **Drift detection** (drift.rs): 96 lines - Desired vs actual state comparison
+- ✅ **Config management** (config.rs): 74 lines - `ConfigMap` creation/updates
+- ✅ **Shared types** (types.rs): 48 lines - Common imports and re-exports
+- ✅ **Total: 1,600 lines** (vs 1,488 original = +112 lines module overhead, 7.5% increase)
+- ✅ **All 594 tests passing** (100% coverage maintained)
+- ✅ **Zero clippy warnings** with pedantic mode
+- ✅ **Backward compatible**: Public API unchanged, all re-exports preserved
+
+---
+
 ## [2026-01-12 19:45] - Phase 4: Records Reconciler Refactoring Complete
 
 **Author:** Erick Bourgeois
@@ -8859,7 +8973,7 @@ let name_server_ips = if spec.name_server_ips.is_none() {
     // Auto-generate from instances
     let mut ordered_instances = primary_instance_refs.clone();
     ordered_instances.extend(secondary_instance_refs.clone());
-    
+
     generate_nameserver_ips(&client, &spec.zone_name, &ordered_instances).await?
 } else {
     // Use explicit user-provided nameserver IPs
