+++ src/reconcilers/bind9cluster/config_tests.rs
@@ -1,0 +1,49 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `config.rs`
+//!
+//! These tests document expected behavior for cluster `ConfigMap` management.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_create_or_update_cluster_configmap_with_custom_refs() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with custom ConfigMap refs (spec.common.configMapRefs)
+        //        AND named_conf is set OR named_conf_options is set
+        // When: create_or_update_cluster_configmap is called
+        // Then: Should return Ok(()) immediately
+        //       AND skip cluster ConfigMap creation
+        //       AND log "uses custom ConfigMaps, skipping cluster ConfigMap creation"
+    }
+
+    #[tokio::test]
+    async fn test_create_or_update_cluster_configmap_creates_new() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster without custom ConfigMap refs
+        //        AND no existing ConfigMap with name "<cluster-name>-config"
+        // When: create_or_update_cluster_configmap is called
+        // Then: Should build cluster ConfigMap from spec.global
+        //       AND create the ConfigMap in Kubernetes
+        //       AND log "Creating cluster ConfigMap <namespace>/<name>"
+    }
+
+    #[tokio::test]
+    async fn test_create_or_update_cluster_configmap_updates_existing() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster without custom ConfigMap refs
+        //        AND existing ConfigMap with name "<cluster-name>-config"
+        // When: create_or_update_cluster_configmap is called
+        // Then: Should build cluster ConfigMap from spec.global
+        //       AND replace the existing ConfigMap in Kubernetes
+        //       AND log "Updating cluster ConfigMap <namespace>/<name>"
+    }
+}
