+++ src/reconcilers/bind9cluster/drift_tests.rs
@@ -1,0 +1,72 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `drift.rs`
+//!
+//! These tests document expected behavior for instance drift detection.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_detect_instance_drift_no_drift() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (2 primary, 3 secondary)
+        //        AND exactly 2 primary instances exist with correct labels
+        //        AND exactly 3 secondary instances exist with correct labels
+        // When: detect_instance_drift is called
+        // Then: Should return Ok(false) - no drift detected
+    }
+
+    #[tokio::test]
+    async fn test_detect_instance_drift_missing_primaries() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (2 primary, 3 secondary)
+        //        AND only 1 primary instance exists
+        //        AND exactly 3 secondary instances exist
+        // When: detect_instance_drift is called
+        // Then: Should return Ok(true) - drift detected
+        //       AND log "Instance drift detected: desired (primary=2, secondary=3), actual (primary=1, secondary=3)"
+    }
+
+    #[tokio::test]
+    async fn test_detect_instance_drift_extra_secondaries() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (2 primary, 3 secondary)
+        //        AND exactly 2 primary instances exist
+        //        AND 5 secondary instances exist (2 extra)
+        // When: detect_instance_drift is called
+        // Then: Should return Ok(true) - drift detected
+        //       AND log "Instance drift detected: desired (primary=2, secondary=3), actual (primary=2, secondary=5)"
+    }
+
+    #[tokio::test]
+    async fn test_detect_instance_drift_filters_by_labels() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster "cluster-a" with desired replicas (2 primary, 0 secondary)
+        //        AND 2 instances with label "bindy.firestoned.io/cluster: cluster-a"
+        //        AND 3 instances with label "bindy.firestoned.io/cluster: cluster-b"
+        // When: detect_instance_drift is called for "cluster-a"
+        // Then: Should only count the 2 instances for cluster-a
+        //       AND return Ok(false) - no drift for cluster-a
+    }
+
+    #[tokio::test]
+    async fn test_detect_instance_drift_zero_desired_replicas() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (0 primary, 0 secondary)
+        //        AND 0 instances exist
+        // When: detect_instance_drift is called
+        // Then: Should return Ok(false) - no drift
+    }
+}
