+++ src/reconcilers/bind9cluster/instances.rs
@@ -1,546 +1,14 @@
 // Copyright (c) 2025 Erick Bourgeois, firestoned
 // SPDX-License-Identifier: MIT
 
-//! BIND9 cluster reconciliation logic.
+//! Instance lifecycle management for `Bind9Cluster` resources.
 //!
-//! This module handles the lifecycle of BIND9 cluster resources in Kubernetes.
-//! It manages the `Bind9Instance` resources that belong to a cluster and updates
-//! the cluster status to reflect the overall health.
+//! This module handles creating, updating, and deleting `Bind9Instance`
+//! resources that are managed by a `Bind9Cluster`.
 
+#[allow(clippy::wildcard_imports)]
+use super::types::*;
 use crate::constants::{API_GROUP_VERSION, KIND_BIND9_CLUSTER, KIND_BIND9_INSTANCE};
-use crate::context::Context;
-use crate::crd::{
-    Bind9Cluster, Bind9ClusterStatus, Bind9Instance, Bind9InstanceSpec, Condition, ServerRole,
-};
-use crate::labels::{
-    BINDY_CLUSTER_LABEL, BINDY_INSTANCE_INDEX_ANNOTATION, BINDY_MANAGED_BY_LABEL,
-    BINDY_RECONCILE_TRIGGER_ANNOTATION, BINDY_ROLE_LABEL, FINALIZER_BIND9_CLUSTER, K8S_PART_OF,
-    MANAGED_BY_BIND9_CLUSTER, PART_OF_BINDY, ROLE_PRIMARY, ROLE_SECONDARY,
-};
-use crate::reconcilers::finalizers::{ensure_finalizer, handle_deletion, FinalizerCleanup};
-use crate::status_reasons::{
-    bind9_instance_condition_type, CONDITION_TYPE_READY, REASON_ALL_READY, REASON_NOT_READY,
-    REASON_NO_CHILDREN, REASON_PARTIALLY_READY, REASON_READY,
-};
-use anyhow::Result;
-use chrono::Utc;
-use k8s_openapi::{
-    api::{
-        apps::v1::Deployment,
-        core::v1::{ConfigMap, Secret, Service},
-    },
-    apimachinery::pkg::apis::meta::v1::ObjectMeta,
-};
-use kube::{
-    api::{DeleteParams, ListParams, Patch, PatchParams, PostParams},
-    client::Client,
-    Api, ResourceExt,
-};
-use serde_json::json;
-use std::collections::BTreeMap;
-use std::sync::Arc;
-use tracing::{debug, error, info, warn};
-
-/// Implement cleanup trait for `Bind9Cluster` finalizer management
-#[async_trait::async_trait]
-impl FinalizerCleanup for Bind9Cluster {
-    async fn cleanup(&self, client: &Client) -> Result<()> {
-        let namespace = self.namespace().unwrap_or_default();
-        let name = self.name_any();
-        delete_cluster_instances(client, &namespace, &name).await
-    }
-}
-
-/// Reconciles a `Bind9Cluster` resource.
-///
-/// This function:
-/// 1. Checks if the cluster is being deleted and handles cleanup
-/// 2. Adds finalizer if not present
-/// 3. Creates/updates cluster `ConfigMap`
-/// 4. Reconciles managed instances
-/// 5. Updates cluster status based on instance health
-///
-/// # Arguments
-///
-/// * `ctx` - Controller context with Kubernetes client and reflector stores
-/// * `cluster` - The `Bind9Cluster` resource to reconcile
-///
-/// # Returns
-///
-/// * `Ok(())` - If reconciliation succeeded
-/// * `Err(_)` - If status update failed
-///
-/// # Errors
-///
-/// Returns an error if Kubernetes API operations fail or status update fails.
-pub async fn reconcile_bind9cluster(ctx: Arc<Context>, cluster: Bind9Cluster) -> Result<()> {
-    let client = ctx.client.clone();
-    let namespace = cluster.namespace().unwrap_or_default();
-    let name = cluster.name_any();
-
-    info!("Reconciling Bind9Cluster: {}/{}", namespace, name);
-    debug!(
-        namespace = %namespace,
-        name = %name,
-        generation = ?cluster.metadata.generation,
-        "Starting Bind9Cluster reconciliation"
-    );
-
-    // Handle deletion if cluster is being deleted
-    if cluster.metadata.deletion_timestamp.is_some() {
-        return handle_deletion(&client, &cluster, FINALIZER_BIND9_CLUSTER).await;
-    }
-
-    // Ensure finalizer is present
-    ensure_finalizer(&client, &cluster, FINALIZER_BIND9_CLUSTER).await?;
-
-    // Check if spec has changed using the standard generation check
-    let current_generation = cluster.metadata.generation;
-    let observed_generation = cluster.status.as_ref().and_then(|s| s.observed_generation);
-
-    // Only reconcile spec-related resources if spec changed OR drift detected
-    let spec_changed =
-        crate::reconcilers::should_reconcile(current_generation, observed_generation);
-
-    // DRIFT DETECTION: Check if managed instances match desired state
-    let drift_detected = if spec_changed {
-        false
-    } else {
-        detect_instance_drift(&client, &cluster, &namespace, &name).await?
-    };
-
-    if spec_changed || drift_detected {
-        if drift_detected {
-            info!(
-                "Spec unchanged but instance drift detected for cluster {}/{}",
-                namespace, name
-            );
-        } else {
-            debug!(
-                "Reconciliation needed: current_generation={:?}, observed_generation={:?}",
-                current_generation, observed_generation
-            );
-        }
-
-        // Create or update shared cluster ConfigMap
-        create_or_update_cluster_configmap(&client, &cluster).await?;
-
-        // Reconcile managed instances (create/update as needed)
-        reconcile_managed_instances(&ctx, &cluster).await?;
-    } else {
-        debug!(
-            "Spec unchanged (generation={:?}) and no drift detected, skipping resource reconciliation",
-            current_generation
-        );
-    }
-
-    // ALWAYS list and analyze cluster instances to update status
-    // This ensures status reflects current instance health even when spec hasn't changed
-    let instances: Vec<Bind9Instance> =
-        list_cluster_instances(&client, &cluster, &namespace, &name).await?;
-
-    // Calculate cluster status from instances
-    let (instance_count, ready_instances, instance_names, conditions) =
-        calculate_cluster_status(&instances, &namespace, &name);
-
-    // Update cluster status with all conditions
-    update_status(
-        &client,
-        &cluster,
-        conditions,
-        instance_count,
-        ready_instances,
-        instance_names,
-    )
-    .await?;
-
-    Ok(())
-}
-
-/// List all `Bind9Instance` resources that reference a cluster
-///
-/// Filters instances in the namespace that have `clusterRef` matching the cluster name.
-///
-/// # Arguments
-///
-/// * `client` - Kubernetes API client
-/// * `cluster` - The `Bind9Cluster` to find instances for
-/// * `namespace` - Cluster namespace
-/// * `name` - Cluster name
-///
-/// # Returns
-///
-/// Vector of `Bind9Instance` resources that reference this cluster
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - Failed to list instances
-/// - Failed to update cluster status on error
-async fn list_cluster_instances(
-    client: &Client,
-    cluster: &Bind9Cluster,
-    namespace: &str,
-    name: &str,
-) -> Result<Vec<Bind9Instance>> {
-    // List all Bind9Instance resources in the namespace that reference this cluster
-    let instances_api: Api<Bind9Instance> = Api::namespaced(client.clone(), namespace);
-    let list_params = ListParams::default();
-    debug!(namespace = %namespace, "Listing Bind9Instance resources");
-
-    match instances_api.list(&list_params).await {
-        Ok(list) => {
-            debug!(
-                total_instances_in_ns = list.items.len(),
-                "Listed Bind9Instance resources"
-            );
-            // Filter instances that reference this cluster
-            let filtered: Vec<_> = list
-                .items
-                .into_iter()
-                .filter(|instance| instance.spec.cluster_ref == name)
-                .collect();
-            debug!(
-                filtered_instances = filtered.len(),
-                cluster_ref = %name,
-                "Filtered instances by cluster reference"
-            );
-            Ok(filtered)
-        }
-        Err(e) => {
-            error!(
-                "Failed to list Bind9Instance resources for cluster {}/{}: {}",
-                namespace, name, e
-            );
-
-            // Update status to show error
-            let error_condition = Condition {
-                r#type: CONDITION_TYPE_READY.to_string(),
-                status: "False".to_string(),
-                reason: Some(REASON_NOT_READY.to_string()),
-                message: Some(format!("Failed to list instances: {e}")),
-                last_transition_time: Some(Utc::now().to_rfc3339()),
-            };
-            update_status(client, cluster, vec![error_condition], 0, 0, vec![]).await?;
-
-            Err(e.into())
-        }
-    }
-}
-
-/// Calculate cluster status from instance health
-///
-/// Analyzes instance list to determine cluster readiness.
-///
-/// # Arguments
-///
-/// * `instances` - List of `Bind9Instance` resources for the cluster
-/// * `namespace` - Cluster namespace (for logging)
-/// * `name` - Cluster name (for logging)
-///
-/// # Returns
-///
-/// Tuple of:
-/// - `instance_count` - Total number of instances
-/// - `ready_instances` - Number of ready instances
-/// - `instance_names` - Names of all instances
-/// - `status` - Cluster status ("True" or "False")
-/// - `message` - Status message
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-pub(crate) fn calculate_cluster_status(
-    instances: &[Bind9Instance],
-    namespace: &str,
-    name: &str,
-) -> (i32, i32, Vec<String>, Vec<Condition>) {
-    // Count total instances and ready instances
-    let instance_count = instances.len() as i32;
-    let instance_names: Vec<String> = instances.iter().map(ResourceExt::name_any).collect();
-
-    let ready_instances = instances
-        .iter()
-        .filter(|instance| {
-            instance
-                .status
-                .as_ref()
-                .and_then(|status| status.conditions.first())
-                .is_some_and(|condition| condition.r#type == "Ready" && condition.status == "True")
-        })
-        .count() as i32;
-
-    info!(
-        "Bind9Cluster {}/{} has {} instances, {} ready",
-        namespace, name, instance_count, ready_instances
-    );
-
-    // Create instance-level conditions
-    let mut instance_conditions = Vec::new();
-    for (index, instance) in instances.iter().enumerate() {
-        let instance_name = instance.name_any();
-        let is_instance_ready = instance
-            .status
-            .as_ref()
-            .and_then(|status| status.conditions.first())
-            .is_some_and(|condition| condition.r#type == "Ready" && condition.status == "True");
-
-        let (status, reason, message) = if is_instance_ready {
-            (
-                "True",
-                REASON_READY,
-                format!("Instance {instance_name} is ready"),
-            )
-        } else {
-            (
-                "False",
-                REASON_NOT_READY,
-                format!("Instance {instance_name} is not ready"),
-            )
-        };
-
-        instance_conditions.push(Condition {
-            r#type: bind9_instance_condition_type(index),
-            status: status.to_string(),
-            reason: Some(reason.to_string()),
-            message: Some(message),
-            last_transition_time: Some(Utc::now().to_rfc3339()),
-        });
-    }
-
-    // Create encompassing Ready condition
-    let (encompassing_status, encompassing_reason, encompassing_message) = if instance_count == 0 {
-        debug!("No instances found for cluster");
-        (
-            "False",
-            REASON_NO_CHILDREN,
-            "No instances found for this cluster".to_string(),
-        )
-    } else if ready_instances == instance_count {
-        debug!("All instances ready");
-        (
-            "True",
-            REASON_ALL_READY,
-            format!("All {instance_count} instances are ready"),
-        )
-    } else if ready_instances > 0 {
-        debug!(ready_instances, instance_count, "Cluster progressing");
-        (
-            "False",
-            REASON_PARTIALLY_READY,
-            format!("{ready_instances}/{instance_count} instances are ready"),
-        )
-    } else {
-        debug!("Waiting for instances to become ready");
-        (
-            "False",
-            REASON_NOT_READY,
-            "No instances are ready".to_string(),
-        )
-    };
-
-    let encompassing_condition = Condition {
-        r#type: CONDITION_TYPE_READY.to_string(),
-        status: encompassing_status.to_string(),
-        reason: Some(encompassing_reason.to_string()),
-        message: Some(encompassing_message.clone()),
-        last_transition_time: Some(Utc::now().to_rfc3339()),
-    };
-
-    // Combine encompassing condition + instance-level conditions
-    let mut all_conditions = vec![encompassing_condition];
-    all_conditions.extend(instance_conditions);
-
-    debug!(
-        status = %encompassing_status,
-        message = %encompassing_message,
-        num_conditions = all_conditions.len(),
-        "Determined cluster status"
-    );
-
-    (
-        instance_count,
-        ready_instances,
-        instance_names,
-        all_conditions,
-    )
-}
-
-/// Update the status of a `Bind9Cluster` with multiple conditions
-async fn update_status(
-    client: &Client,
-    cluster: &Bind9Cluster,
-    conditions: Vec<Condition>,
-    instance_count: i32,
-    ready_instances: i32,
-    instances: Vec<String>,
-) -> Result<()> {
-    let api: Api<Bind9Cluster> =
-        Api::namespaced(client.clone(), &cluster.namespace().unwrap_or_default());
-
-    // Check if status has actually changed
-    let current_status = &cluster.status;
-    let status_changed =
-        if let Some(current) = current_status {
-            // Check if counts changed
-            if current.instance_count != Some(instance_count)
-                || current.ready_instances != Some(ready_instances)
-                || current.instances != instances
-            {
-                true
-            } else {
-                // Check if any condition changed
-                if current.conditions.len() == conditions.len() {
-                    // Compare each condition
-                    current.conditions.iter().zip(conditions.iter()).any(
-                        |(current_cond, new_cond)| {
-                            current_cond.r#type != new_cond.r#type
-                                || current_cond.status != new_cond.status
-                                || current_cond.message != new_cond.message
-                                || current_cond.reason != new_cond.reason
-                        },
-                    )
-                } else {
-                    true
-                }
-            }
-        } else {
-            // No status exists, need to update
-            true
-        };
-
-    // Only update if status has changed
-    if !status_changed {
-        debug!(
-            namespace = %cluster.namespace().unwrap_or_default(),
-            name = %cluster.name_any(),
-            "Status unchanged, skipping update"
-        );
-        info!(
-            "Bind9Cluster {}/{} status unchanged, skipping update",
-            cluster.namespace().unwrap_or_default(),
-            cluster.name_any()
-        );
-        return Ok(());
-    }
-
-    debug!(
-        instance_count,
-        ready_instances,
-        instances_count = instances.len(),
-        num_conditions = conditions.len(),
-        "Preparing status update"
-    );
-
-    let new_status = Bind9ClusterStatus {
-        conditions,
-        observed_generation: cluster.metadata.generation,
-        instance_count: Some(instance_count),
-        ready_instances: Some(ready_instances),
-        instances,
-    };
-
-    info!(
-        "Updating Bind9Cluster {}/{} status: {} instances, {} ready",
-        cluster.namespace().unwrap_or_default(),
-        cluster.name_any(),
-        instance_count,
-        ready_instances
-    );
-
-    let patch = json!({ "status": new_status });
-    api.patch_status(
-        &cluster.name_any(),
-        &PatchParams::apply("bindy-controller"),
-        &Patch::Merge(&patch),
-    )
-    .await?;
-
-    Ok(())
-}
-
-/// Detects if the actual managed instances match the desired replica counts.
-///
-/// Compares the number of primary and secondary instances that exist against
-/// the desired replica counts in the cluster spec.
-///
-/// # Arguments
-///
-/// * `client` - Kubernetes API client
-/// * `cluster` - The `Bind9Cluster` to check
-/// * `namespace` - Cluster namespace
-/// * `name` - Cluster name
-///
-/// # Returns
-///
-/// * `Ok(true)` - Drift detected (instances don't match desired state)
-/// * `Ok(false)` - No drift (instances match desired state)
-/// * `Err(_)` - Failed to check drift
-///
-/// # Errors
-///
-/// Returns an error if listing instances fails.
-async fn detect_instance_drift(
-    client: &Client,
-    cluster: &Bind9Cluster,
-    namespace: &str,
-    name: &str,
-) -> Result<bool> {
-    // Get desired replica counts from spec
-    let desired_primary = cluster
-        .spec
-        .common
-        .primary
-        .as_ref()
-        .and_then(|p| p.replicas)
-        .unwrap_or(0);
-
-    let desired_secondary = cluster
-        .spec
-        .common
-        .secondary
-        .as_ref()
-        .and_then(|s| s.replicas)
-        .unwrap_or(0);
-
-    // List existing managed instances
-    let api: Api<Bind9Instance> = Api::namespaced(client.clone(), namespace);
-    let instances = api.list(&ListParams::default()).await?;
-
-    // Filter for managed instances of this cluster
-    let managed_instances: Vec<_> = instances
-        .items
-        .into_iter()
-        .filter(|instance| {
-            instance.metadata.labels.as_ref().is_some_and(|labels| {
-                labels.get(BINDY_MANAGED_BY_LABEL) == Some(&MANAGED_BY_BIND9_CLUSTER.to_string())
-                    && labels.get(BINDY_CLUSTER_LABEL) == Some(&name.to_string())
-            })
-        })
-        .collect();
-
-    // Count by role
-    let actual_primary = managed_instances
-        .iter()
-        .filter(|i| i.spec.role == ServerRole::Primary)
-        .count();
-
-    let actual_secondary = managed_instances
-        .iter()
-        .filter(|i| i.spec.role == ServerRole::Secondary)
-        .count();
-
-    // Drift detected if counts don't match
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    let drift = actual_primary != desired_primary as usize
-        || actual_secondary != desired_secondary as usize;
-
-    if drift {
-        info!(
-            "Instance drift detected for cluster {}/{}: desired (primary={}, secondary={}), actual (primary={}, secondary={})",
-            namespace, name, desired_primary, desired_secondary, actual_primary, actual_secondary
-        );
-    }
-
-    Ok(drift)
-}
 
 /// Reconcile managed `Bind9Instance` resources for a cluster
 ///
@@ -558,7 +26,10 @@ async fn detect_instance_drift(
 /// - Failed to list existing instances
 /// - Failed to create new instances
 #[allow(clippy::too_many_lines)]
-async fn reconcile_managed_instances(ctx: &Context, cluster: &Bind9Cluster) -> Result<()> {
+pub(super) async fn reconcile_managed_instances(
+    ctx: &Context,
+    cluster: &Bind9Cluster,
+) -> Result<()> {
     let client = ctx.client.clone();
     let namespace = cluster.namespace().unwrap_or_default();
     let cluster_name = cluster.name_any();
@@ -849,7 +320,7 @@ async fn reconcile_managed_instances(ctx: &Context, cluster: &Bind9Cluster) -> R
 /// # Errors
 ///
 /// Returns an error if patching instances fails
-async fn update_existing_managed_instances(
+pub(super) async fn update_existing_managed_instances(
     client: &Client,
     namespace: &str,
     cluster_name: &str,
@@ -970,7 +441,7 @@ async fn update_existing_managed_instances(
 /// # Errors
 ///
 /// Returns an error if resource checking or instance update fails
-async fn ensure_managed_instance_resources(
+pub(super) async fn ensure_managed_instance_resources(
     client: &Client,
     cluster: &Bind9Cluster,
     managed_instances: &[Bind9Instance],
@@ -1280,65 +751,7 @@ async fn create_managed_instance_with_owner(
     }
 }
 
-/// Create or update the shared cluster-level `ConfigMap`
-///
-/// This `ConfigMap` contains BIND9 configuration that is shared across all instances
-/// in the cluster. It is created from `spec.global` configuration.
-///
-/// # Arguments
-///
-/// * `client` - Kubernetes API client
-/// * `cluster` - The `Bind9Cluster` resource
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - Failed to create or update the `ConfigMap`
-/// - Kubernetes API operations fail
-async fn create_or_update_cluster_configmap(client: &Client, cluster: &Bind9Cluster) -> Result<()> {
-    use crate::bind9_resources::build_cluster_configmap;
-
-    let namespace = cluster.namespace().unwrap_or_default();
-    let name = cluster.name_any();
-
-    // Check if custom ConfigMaps are referenced at the cluster level
-    if let Some(refs) = &cluster.spec.common.config_map_refs {
-        if refs.named_conf.is_some() || refs.named_conf_options.is_some() {
-            info!(
-                "Cluster {}/{} uses custom ConfigMaps, skipping cluster ConfigMap creation",
-                namespace, name
-            );
-            return Ok(());
-        }
-    }
-
-    info!(
-        "Creating/updating shared ConfigMap for cluster {}/{}",
-        namespace, name
-    );
-
-    // Build the cluster ConfigMap
-    let configmap = build_cluster_configmap(&name, &namespace, cluster)?;
-
-    let cm_api: Api<ConfigMap> = Api::namespaced(client.clone(), &namespace);
-    let cm_name = format!("{name}-config");
-
-    if (cm_api.get(&cm_name).await).is_ok() {
-        // ConfigMap exists, update it
-        info!("Updating cluster ConfigMap {}/{}", namespace, cm_name);
-        cm_api
-            .replace(&cm_name, &PostParams::default(), &configmap)
-            .await?;
-    } else {
-        // ConfigMap doesn't exist, create it
-        info!("Creating cluster ConfigMap {}/{}", namespace, cm_name);
-        cm_api.create(&PostParams::default(), &configmap).await?;
-    }
-
-    Ok(())
-}
-
-/// Delete a single managed `Bind9Instance` resource
+/// Delete a single managed `Bind9Instance` resource.
 ///
 /// This function is public to allow reuse by `ClusterBind9Provider` reconciler.
 ///
@@ -1396,7 +809,7 @@ pub async fn delete_managed_instance(
 /// Returns an error if:
 /// - Failed to list `Bind9Instance` resources
 /// - Failed to delete any `Bind9Instance` resource
-async fn delete_cluster_instances(
+pub(super) async fn delete_cluster_instances(
     client: &Client,
     namespace: &str,
     cluster_name: &str,
@@ -1486,3 +899,7 @@ pub async fn delete_bind9cluster(_client: Client, _cluster: Bind9Cluster) -> Res
     // Deletion is now handled by the finalizer in reconcile_bind9cluster
     Ok(())
 }
+
+#[cfg(test)]
+#[path = "instances_tests.rs"]
+mod instances_tests;
