+++ src/reconcilers/bind9cluster/instances_tests.rs
@@ -1,0 +1,144 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `instances.rs`
+//!
+//! These tests document expected behavior for instance lifecycle management.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    #[tokio::test]
+    async fn test_reconcile_managed_instances_zero_replicas() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with spec.common.primary.replicas = 0
+        //        AND spec.common.secondary.replicas = 0
+        // When: reconcile_managed_instances is called
+        // Then: Should return Ok(()) immediately
+        //       AND log "No instances requested for cluster"
+        //       AND NOT create any instances
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_managed_instances_scale_up() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (2 primary, 1 secondary)
+        //        AND 0 existing managed instances
+        // When: reconcile_managed_instances is called
+        // Then: Should create 2 primary instances (cluster-name-primary-0, cluster-name-primary-1)
+        //       AND create 1 secondary instance (cluster-name-secondary-0)
+        //       AND log "Scaled cluster: created 2 primary, 1 secondary; deleted 0 primary, 0 secondary"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_managed_instances_scale_down() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (1 primary, 1 secondary)
+        //        AND 3 existing primary instances (indexes 0, 1, 2)
+        //        AND 2 existing secondary instances (indexes 0, 1)
+        // When: reconcile_managed_instances is called
+        // Then: Should delete highest-indexed primaries (cluster-name-primary-2, cluster-name-primary-1)
+        //       AND delete highest-indexed secondary (cluster-name-secondary-1)
+        //       AND log "Scaled cluster: created 0 primary, 0 secondary; deleted 2 primary, 1 secondary"
+    }
+
+    #[tokio::test]
+    async fn test_reconcile_managed_instances_no_drift() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with desired replicas (2 primary, 3 secondary)
+        //        AND exactly 2 primary instances exist
+        //        AND exactly 3 secondary instances exist
+        // When: reconcile_managed_instances is called
+        // Then: Should NOT create or delete any instances
+        //       AND log "Cluster already at desired scale"
+    }
+
+    #[tokio::test]
+    async fn test_update_existing_managed_instances_no_drift() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 managed instances exist
+        //        AND all instances match cluster spec (version, image, configMapRefs, etc.)
+        // When: update_existing_managed_instances is called
+        // Then: Should NOT patch any instances
+        //       AND log "Instance <name> spec matches cluster spec, no update needed" for each
+    }
+
+    #[tokio::test]
+    async fn test_update_existing_managed_instances_version_drift() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 2 managed instances exist
+        //        AND cluster spec has version="9.18.24"
+        //        AND instances have version="9.18.20"
+        // When: update_existing_managed_instances is called
+        // Then: Should patch both instances with updated version
+        //       AND log "Instance <name> spec differs from cluster spec, updating"
+        //       AND log "Updated 2 managed instances in cluster to match current spec"
+    }
+
+    #[tokio::test]
+    async fn test_ensure_managed_instance_resources_all_exist() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 2 managed instances
+        //        AND all resources exist (ConfigMap, Secret, Service, Deployment)
+        // When: ensure_managed_instance_resources is called
+        // Then: Should NOT trigger reconciliation for any instance
+        //       AND log "All child resources exist for managed instance" for each
+    }
+
+    #[tokio::test]
+    async fn test_ensure_managed_instance_resources_missing_deployment() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 1 managed instance
+        //        AND ConfigMap, Secret, Service exist
+        //        AND Deployment is missing
+        // When: ensure_managed_instance_resources is called
+        // Then: Should patch instance with reconcile trigger annotation
+        //       AND log "Missing resources for managed instance: Deployment. Triggering reconciliation."
+        //       AND log "Triggered reconciliation for instance to recreate: Deployment"
+    }
+
+    #[tokio::test]
+    async fn test_create_managed_instance_with_labels() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with spec.common.primary.labels = {env: prod, team: platform}
+        // When: create_managed_instance is called for role=Primary
+        // Then: Should create instance with labels including:
+        //       - bindy.firestoned.io/managed-by: bind9cluster
+        //       - bindy.firestoned.io/cluster: <cluster-name>
+        //       - bindy.firestoned.io/role: primary
+        //       - env: prod (propagated from cluster)
+        //       - team: platform (propagated from cluster)
+    }
+
+    #[tokio::test]
+    async fn test_delete_cluster_instances_multiple_instances() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 5 Bind9Instance resources in namespace
+        //        AND 3 have spec.clusterRef = "cluster-a"
+        //        AND 2 have spec.clusterRef = "cluster-b"
+        // When: delete_cluster_instances is called for "cluster-a"
+        // Then: Should delete only the 3 instances with clusterRef="cluster-a"
+        //       AND log "Found 3 Bind9Instance resources for cluster, deleting..."
+        //       AND log "Successfully deleted all Bind9Instance resources for cluster"
+    }
+}
