+++ src/reconcilers/dnszone/bind9_config_tests.rs
@@ -1,0 +1,38 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `bind9_config.rs`
+//!
+//! These tests document expected behavior for BIND9 configuration orchestration.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::InstanceReference;
+
+    #[tokio::test]
+    async fn test_configure_zone_requires_primary_instances() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A zone with no primary instances
+        // When: configure_zone_on_instances is called
+        // Then: Should return error "No primary servers found"
+        //       AND set Degraded condition with reason "PrimaryFailed"
+    }
+
+    #[test]
+    fn test_instance_reference_construction() {
+        let instance_ref = InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: "test-instance".to_string(),
+            namespace: "default".to_string(),
+            last_reconciled_at: None,
+        };
+
+        assert_eq!(instance_ref.name, "test-instance");
+        assert_eq!(instance_ref.namespace, "default");
+        assert!(instance_ref.last_reconciled_at.is_none());
+    }
+}
