+++ src/reconcilers/dnszone/cleanup_tests.rs
@@ -1,0 +1,45 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `cleanup.rs`
+//!
+//! These tests document expected behavior for cleanup operations.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::InstanceReference;
+
+    #[tokio::test]
+    async fn test_cleanup_deleted_instances_no_instances() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A zone with no instances in status
+        // When: cleanup_deleted_instances is called
+        // Then: Should return Ok(0) immediately
+        //       AND log debug message "No instances in status - skipping cleanup"
+    }
+
+    #[test]
+    fn test_instance_reference_equality() {
+        let inst1 = InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: "instance-1".to_string(),
+            namespace: "default".to_string(),
+            last_reconciled_at: None,
+        };
+
+        let inst2 = InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: "instance-1".to_string(),
+            namespace: "default".to_string(),
+            last_reconciled_at: Some("2025-01-01T00:00:00Z".to_string()),
+        };
+
+        // InstanceReference equality should ignore last_reconciled_at
+        assert_eq!(inst1, inst2);
+    }
+}
