+++ src/reconcilers/dnszone/discovery_tests.rs
@@ -1,0 +1,56 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `discovery.rs`
+//!
+//! These tests document expected behavior for DNS record discovery.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::LabelSelector;
+    use std::collections::BTreeMap;
+
+    #[tokio::test]
+    async fn test_reconcile_zone_records_no_selectors() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A zone with no recordsFrom selectors
+        // When: reconcile_zone_records is called
+        // Then: Should return Ok(vec![]) immediately
+        //       AND log "No label selectors defined - skipping record discovery"
+    }
+
+    #[test]
+    fn test_label_selector_matches_exact() {
+        let mut labels = BTreeMap::new();
+        labels.insert("app".to_string(), "nginx".to_string());
+        labels.insert("env".to_string(), "prod".to_string());
+
+        let selector = LabelSelector {
+            match_labels: Some(labels.clone()),
+            match_expressions: None,
+        };
+
+        // Should match exact labels
+        assert!(selector.matches(&labels));
+    }
+
+    #[test]
+    fn test_label_selector_no_match_different_value() {
+        let mut selector_labels = BTreeMap::new();
+        selector_labels.insert("app".to_string(), "nginx".to_string());
+
+        let mut pod_labels = BTreeMap::new();
+        pod_labels.insert("app".to_string(), "apache".to_string());
+
+        let selector = LabelSelector {
+            match_labels: Some(selector_labels),
+            match_expressions: None,
+        };
+
+        // Should not match when value differs
+        assert!(!selector.matches(&pod_labels));
+    }
+}
