+++ src/reconcilers/dnszone/secondary_tests.rs
@@ -1,0 +1,166 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `secondary.rs`
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::InstanceReference;
+
+    /// Helper to create test instance references
+    fn create_instance_ref(name: &str, namespace: &str) -> InstanceReference {
+        InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: name.to_string(),
+            namespace: namespace.to_string(),
+            last_reconciled_at: None,
+        }
+    }
+
+    #[tokio::test]
+    async fn test_filter_secondary_instances_all_secondary() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 instance references
+        //        AND all 3 instances have role=Secondary
+        // When: filter_secondary_instances is called
+        // Then: Should return all 3 instance references
+    }
+
+    #[tokio::test]
+    async fn test_filter_secondary_instances_mixed_roles() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 5 instance references
+        //        AND 3 instances have role=Secondary
+        //        AND 2 instances have role=Primary
+        // When: filter_secondary_instances is called
+        // Then: Should return only the 3 secondary instance references
+    }
+
+    #[tokio::test]
+    async fn test_filter_secondary_instances_none_secondary() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 instance references
+        //        AND all 3 instances have role=Primary
+        // When: filter_secondary_instances is called
+        // Then: Should return empty vec
+    }
+
+    #[tokio::test]
+    async fn test_find_secondary_pod_ips_from_instances() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 4 instance references
+        //        AND 2 are SECONDARY instances
+        //        AND secondary-1 has 1 running pod (IP: 10.0.1.1)
+        //        AND secondary-2 has 2 running pods (IPs: 10.0.2.1, 10.0.2.2)
+        // When: find_secondary_pod_ips_from_instances is called
+        // Then: Should return vec!["10.0.1.1", "10.0.2.1", "10.0.2.2"]
+    }
+
+    #[tokio::test]
+    async fn test_find_secondary_pod_ips_skips_primary() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 instance references
+        //        AND 1 is SECONDARY with 1 running pod (IP: 10.0.1.1)
+        //        AND 2 are PRIMARY
+        // When: find_secondary_pod_ips_from_instances is called
+        // Then: Should return vec!["10.0.1.1"]
+        //       AND not include primary pod IPs
+        //       AND log debug "Skipping instance ... - role is Primary, not Secondary"
+    }
+
+    #[tokio::test]
+    async fn test_find_secondary_pod_ips_skips_non_running() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 1 SECONDARY instance with 3 pods
+        //        AND 2 pods in Running phase
+        //        AND 1 pod in Terminating phase
+        // When: find_secondary_pod_ips_from_instances is called
+        // Then: Should return only the 2 running pod IPs
+        //       AND log debug "Skipping pod ... in phase Terminating"
+    }
+
+    #[tokio::test]
+    async fn test_find_secondary_pod_ips_handles_missing_instances() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 instance references
+        //        AND 1 instance no longer exists in the API
+        //        AND 2 instances are SECONDARY
+        // When: find_secondary_pod_ips_from_instances is called
+        // Then: Should skip the missing instance
+        //       AND return IPs from the 2 existing secondary instances
+        //       AND log warning "Failed to get Bind9Instance ... Skipping"
+    }
+
+    #[tokio::test]
+    async fn test_for_each_secondary_endpoint_success() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with 2 SECONDARY instances
+        //        AND each instance has 1 endpoint
+        // When: for_each_secondary_endpoint is called with an operation
+        // Then: Should execute the operation on both endpoints
+        //       AND return Ok((Some(first_endpoint), 2))
+    }
+
+    #[tokio::test]
+    async fn test_for_each_secondary_endpoint_with_rndc_key() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with 2 SECONDARY instances
+        //        AND each instance has RNDC key secret
+        // When: for_each_secondary_endpoint is called with with_rndc_key=true
+        // Then: Should load RNDC key for each instance
+        //       AND pass key_data to the operation closure
+    }
+
+    #[tokio::test]
+    async fn test_for_each_secondary_endpoint_continues_on_error() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with 3 SECONDARY instances
+        //        AND the operation fails on the 2nd instance
+        // When: for_each_secondary_endpoint is called
+        // Then: Should continue processing the 3rd instance
+        //       AND return error with all failures listed
+        //       AND log error "Failed operation on secondary endpoint ..."
+    }
+
+    #[tokio::test]
+    async fn test_for_each_secondary_endpoint_no_secondaries() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: A cluster with no SECONDARY instances
+        // When: for_each_secondary_endpoint is called
+        // Then: Should return Ok((None, 0))
+        //       AND log "Found 0 SECONDARY pod(s) for cluster"
+    }
+
+    #[test]
+    fn test_instance_reference_secondary_identity() {
+        let ref1 = create_instance_ref("secondary-1", "default");
+        let ref2 = create_instance_ref("secondary-1", "default");
+        let ref3 = create_instance_ref("secondary-2", "default");
+
+        assert_eq!(ref1, ref2);
+        assert_ne!(ref1, ref3);
+    }
+}
