+++ src/reconcilers/dnszone/status_helpers_tests.rs
@@ -1,0 +1,123 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `status_helpers.rs`
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::InstanceReference;
+
+    /// Helper to create test instance references
+    fn create_instance_ref(name: &str, namespace: &str) -> InstanceReference {
+        InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: name.to_string(),
+            namespace: namespace.to_string(),
+            last_reconciled_at: None,
+        }
+    }
+
+    #[tokio::test]
+    async fn test_calculate_expected_instance_counts() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 5 instance references
+        //        AND 2 are PRIMARY
+        //        AND 3 are SECONDARY
+        // When: calculate_expected_instance_counts is called
+        // Then: Should return (2, 3)
+    }
+
+    #[tokio::test]
+    async fn test_calculate_expected_instance_counts_no_secondaries() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: 3 instance references
+        //        AND all 3 are PRIMARY
+        //        AND 0 are SECONDARY
+        // When: calculate_expected_instance_counts is called
+        // Then: Should return (3, 0)
+    }
+
+    #[tokio::test]
+    async fn test_finalize_zone_status_all_ready() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: No degraded conditions
+        //        AND primary_count == expected_primary_count
+        //        AND secondary_count == expected_secondary_count
+        // When: finalize_zone_status is called
+        // Then: Should set Ready=True condition
+        //       AND clear any stale Degraded condition
+        //       AND apply status to API
+    }
+
+    #[tokio::test]
+    async fn test_finalize_zone_status_degraded_condition_exists() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: status_updater has a Degraded condition set
+        //        AND primary_count == expected_primary_count
+        // When: finalize_zone_status is called
+        // Then: Should keep the Degraded condition
+        //       AND NOT overwrite with Ready condition
+        //       AND log "reconciliation completed with degraded state"
+    }
+
+    #[tokio::test]
+    async fn test_finalize_zone_status_partial_reconciliation() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: No degraded conditions
+        //        AND primary_count = 1, expected_primary_count = 2
+        //        AND secondary_count = 2, expected_secondary_count = 3
+        // When: finalize_zone_status is called
+        // Then: Should set Degraded=True condition
+        //       AND reason should be "PartialReconciliation"
+        //       AND message should include "1 instance(s) pending"
+        //       AND log "partially configured: 1/2 primaries, 2/3 secondaries"
+    }
+
+    #[tokio::test]
+    async fn test_finalize_zone_status_partial_primary_only() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: No degraded conditions
+        //        AND primary_count = 1, expected_primary_count = 3
+        //        AND secondary_count = 0, expected_secondary_count = 0
+        // When: finalize_zone_status is called
+        // Then: Should set Degraded=True condition
+        //       AND reason should be "PartialReconciliation"
+        //       AND message should include "2 instance(s) pending"
+    }
+
+    #[tokio::test]
+    async fn test_finalize_zone_status_sets_observed_generation() {
+        // This test requires mocking the Kubernetes API
+        // For now, we document the expected behavior:
+        //
+        // Given: generation = Some(5)
+        // When: finalize_zone_status is called
+        // Then: Should call status_updater.set_observed_generation(Some(5))
+        //       AND observedGeneration should be set in status
+    }
+
+    #[test]
+    fn test_instance_reference_counts() {
+        // Test that we can count instance references
+        let refs = [
+            create_instance_ref("instance-1", "default"),
+            create_instance_ref("instance-2", "default"),
+            create_instance_ref("instance-3", "other"),
+        ];
+
+        assert_eq!(refs.len(), 3);
+    }
+}
