+++ src/reconcilers/dnszone/validation_tests.rs
@@ -1,0 +1,71 @@
+// Copyright (c) 2025 Erick Bourgeois, firestoned
+// SPDX-License-Identifier: MIT
+
+//! Unit tests for `validation.rs`
+//!
+//! These tests document expected behavior for validation logic.
+//! Full implementation requires Kubernetes API mocking infrastructure.
+
+#[cfg(test)]
+mod tests {
+    use crate::crd::InstanceReference;
+    use crate::reconcilers::dnszone::validation::filter_instances_needing_reconciliation;
+
+    fn create_instance_ref(name: &str, namespace: &str) -> InstanceReference {
+        InstanceReference {
+            api_version: "bindy.firestoned.io/v1beta1".to_string(),
+            kind: "Bind9Instance".to_string(),
+            name: name.to_string(),
+            namespace: namespace.to_string(),
+            last_reconciled_at: None,
+        }
+    }
+
+    #[test]
+    fn test_filter_instances_needing_reconciliation_all_need_reconciliation() {
+        let instances = vec![
+            create_instance_ref("instance-1", "default"),
+            create_instance_ref("instance-2", "default"),
+            create_instance_ref("instance-3", "default"),
+        ];
+
+        let result = filter_instances_needing_reconciliation(&instances);
+
+        assert_eq!(result.len(), 3);
+    }
+
+    #[test]
+    fn test_filter_instances_needing_reconciliation_some_already_reconciled() {
+        let mut instances = vec![
+            create_instance_ref("instance-1", "default"),
+            create_instance_ref("instance-2", "default"),
+            create_instance_ref("instance-3", "default"),
+        ];
+
+        // Set timestamp on instance-2 (already reconciled)
+        instances[1].last_reconciled_at = Some("2025-01-01T00:00:00Z".to_string());
+
+        let result = filter_instances_needing_reconciliation(&instances);
+
+        // Should only return instance-1 and instance-3
+        assert_eq!(result.len(), 2);
+        assert_eq!(result[0].name, "instance-1");
+        assert_eq!(result[1].name, "instance-3");
+    }
+
+    #[test]
+    fn test_filter_instances_needing_reconciliation_none_need_reconciliation() {
+        let mut instances = vec![
+            create_instance_ref("instance-1", "default"),
+            create_instance_ref("instance-2", "default"),
+        ];
+
+        // Set timestamp on all instances (all already reconciled)
+        instances[0].last_reconciled_at = Some("2025-01-01T00:00:00Z".to_string());
+        instances[1].last_reconciled_at = Some("2025-01-01T00:00:01Z".to_string());
+
+        let result = filter_instances_needing_reconciliation(&instances);
+
+        assert_eq!(result.len(), 0);
+    }
+}
